@@ -2,10 +2,11 @@
   "name": "next-ts-template",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
     "404": "cp out/404/index.html out/404.html",
     "test": "jest --passWithNoTests",
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "prebuild": "pnpm install",
     "build": "next build",
     "start": "next start",
@@ -26,31 +27,32 @@
     "@emotion/react": "^11.14.0",
     "@emotion/server": "^11.11.0",
     "@emotion/styled": "^11.14.0",
-    "@mui/material": "^5.17.1",
-    "@mui/material-nextjs": "^6.4.3",
+    "@mui/material": "^7.0.2",
+    "@mui/material-nextjs": "^7.0.2",
     "i18next": "^23.16.8",
-    "next": "^15.2.4",
+    "next": "^15.3.0",
     "next-i18next": "^15.4.2",
     "next-language-detector": "^1.1.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-i18next": "^14.1.3",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-i18next": "^15.4.1",
     "web-vitals": "^3.5.2"
   },
   "devDependencies": {
-    "@mui/icons-material": "^5.17.1",
-    "@next/bundle-analyzer": "^14.2.26",
+    "@mui/icons-material": "^7.0.2",
+    "@next/bundle-analyzer": "^15.3.0",
     "@swc/jest": "^0.2.37",
+    "@tailwindcss/postcss": "^4.1.3",
     "@types/jest": "^29.5.14",
     "@types/node": "^20.17.30",
-    "@types/react": "^18.3.20",
+    "@types/react": "^19.1.0",
     "autoprefixer": "^10.4.21",
     "http-server": "^14.1.1",
     "jest": "^29.7.0",
     "postcss": "^8.5.3",
-    "rimraf": "^5.0.10",
+    "rimraf": "^6.0.1",
     "styled-jsx": "^5.1.6",
-    "tailwindcss": "^3.4.17",
+    "tailwindcss": "^4.1.3",
     "typescript": "^5.8.3"
   },
   "jest": {
@@ -64,5 +66,10 @@
     "npm": "use pnpm please!",
     "yarn": "use pnpm please!",
     "pnpm": ">=8.6.12"
+  },
+  "pnpm": {
+    "overrides": {
+      "@types/react": "19.1.0"
+    }
   }
 }
