@@ -1,6 +1,5 @@
-module.exports = {
-    plugins: {
-        tailwindcss: {},
-        autoprefixer: {},
-    },
+export default {
+  plugins: {
+    "@tailwindcss/postcss": {},
+  },
 };
